Revert commit

+++ srcs/ft_getlinesize.c
@@ -3,18 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   ft_getlinesize.c                                   :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: nmougino <nmougino@student.42.fr>          +#+  +:+       +#+        */
+/*   By: nmougino <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2015/12/09 05:13:05 by nmougino          #+#    #+#             */
-/*   Updated: 2016/09/29 01:06:15 by nmougino         ###   ########.fr       */
+/*   Updated: 2015/12/09 12:12:40 by nmougino         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft.h"
-
-size_t	ft_getlinesize(char *buf)
+int		ft_getlinesize(char *buf)
 {
-	size_t	ans;
+	int		ans;
 
 	ans = 0;
 	while (*buf && *buf != '\n')
