+++ srcs/ft_sitoa.c
@@ -1,18 +1,18 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_sitoa.c                                         :+:      :+:    :+:   */
+/*   sitoa.c                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: nmougino <nmougino@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/06/12 19:11:09 by nmougino          #+#    #+#             */
-/*   Updated: 2016/09/29 02:44:01 by nmougino         ###   ########.fr       */
+/*   Updated: 2016/06/12 19:11:10 by nmougino         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	sitoa(intmax_t p, t_print *print, size_t l)
+void	sitoa(intmax_t p, t_print *print, int l)
 {
 	char	c[l + 1];
