+++ srcs/ft_uitoabase.c
@@ -1,18 +1,18 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_uitoabase.c                                     :+:      :+:    :+:   */
+/*   uitoabase.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: nmougino <nmougino@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/06/12 19:11:12 by nmougino          #+#    #+#             */
-/*   Updated: 2016/09/29 02:47:22 by nmougino         ###   ########.fr       */
+/*   Updated: 2016/06/12 22:14:31 by nmougino         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	uitoabase(uintmax_t p, char *base, t_print *print, size_t l)
+void	uitoabase(uintmax_t p, char *base, t_print *print, int l)
 {
 	char	c[l + 1];
 	char	baselen;
